refactor: :recycle: improve folder structure, fix some styling, add responsiveness to CityAutoCompleteInput
